+++ CHANGELOG.md
@@ -10,6 +10,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 ## [Unreleased]
 
 ### Added
+- Using Spring Boot _1.5.2.RELEASE_
 - Added logging formatter/anonymizer for email, used by default in default email scheduler and sender.
 - `SchedulerService` has two new methods that schedule an email at time now with time zone UTC.